+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -97,6 +97,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Added constructors.
   - Added getters.
   - Added mutation methods.
+* Added simple `#define`ed constants.
 
 ### Changed
 * Made `CFArray`, `CFBag`, `CFBinaryHeap`, `CFDictionary`, `CFSet` and their